requirements.txt
accelerate==1.0.1
deepface==0.0.93
diffusers==0.25.1
dlib==19.24.6
face-alignment==1.4.1
face-recognition==1.3.0
huggingface_hub==0.25.0
tf-keras==2.19.0
pytorch-fid==0.3.0
transformers==4.48.0